InitialCommit

+++ src/main/java/com/apsn/MarineClinic/dto/response/UserResponse.java
@@ -2,5 +2,5 @@ package com.apsn.MarineClinic.dto.response;
 
 import java.time.LocalDateTime;
 
-public record UserResponse(Integer user_Id, String name, String email, String password, LocalDateTime createdAt, LocalDateTime updatedAt){
+public record UserResponse(Long user_Id, String name, String email, String password, LocalDateTime createdAt, LocalDateTime updatedAt){
 }
